fix: fix build issues
