+++ packages/ceramic-cli/src/__tests__/ceramic-daemon.test.ts
@@ -4,14 +4,41 @@ import IdentityWallet from 'identity-wallet'
 import tmp from 'tmp-promise'
 import Ipfs from 'ipfs'
 import CeramicDaemon from '../ceramic-daemon'
-import { AnchorStatus, IpfsUtils } from "@ceramicnetwork/ceramic-common"
+import { AnchorStatus } from "@ceramicnetwork/ceramic-common"
 import { TileDoctypeHandler } from "@ceramicnetwork/ceramic-doctype-tile"
 import { EventEmitter } from "events"
+
+import dagJose from 'dag-jose'
+import basicsImport from 'multiformats/cjs/src/basics-import.js'
+import legacy from 'multiformats/cjs/src/legacy.js'
 
 const seed = '0x5872d6e0ae7347b72c9216db218ebbb9d9d0ae7ab818ead3557e8e78bf944184'
 const anchorUpdate = (doc: EventEmitter): Promise<void> => new Promise(resolve => doc.on('change', resolve))
 const port = 7777
 const apiUrl = 'http://localhost:' + port
+
+/**
+ * Create an IPFS instance
+ * @param overrideConfig - IFPS config for override
+ */
+const createIPFS =(overrideConfig: object = {}): Promise<any> => {
+  basicsImport.multicodec.add(dagJose)
+  const format = legacy(basicsImport, dagJose.name)
+
+  const config = {
+    ipld: { formats: [format] },
+    libp2p: {
+      config: {
+        dht: {
+          enabled: true
+        }
+      }
+    }
+  }
+
+  Object.assign(config, overrideConfig)
+  return Ipfs.create(config)
+}
 
 describe('Ceramic interop: core <> http-client', () => {
   jest.setTimeout(20000)
@@ -25,7 +52,7 @@ describe('Ceramic interop: core <> http-client', () => {
 
   beforeAll(async () => {
     tmpFolder = await tmp.dir({ unsafeCleanup: true })
-    ipfs = await IpfsUtils.createIPFS({
+    ipfs = await createIPFS({
       repo: `${tmpFolder.path}/ipfs${7}/`,
       config: {
         Addresses: { Swarm: [ `/ip4/127.0.0.1/tcp/${4011}` ] },
