+++ packages/ceramic-common/package.json
@@ -32,8 +32,7 @@
     "dag-jose": "^0.1.1",
     "did-resolver": "^2.1.1",
     "ipfs": "^0.49.0",
-    "lodash.clonedeep": "^4.5.0",
-    "multiformats": "^3.0.3"
+    "lodash.clonedeep": "^4.5.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
