+++ packages/ceramic-common/src/index.ts
@@ -3,4 +3,3 @@ export * from './ceramic-api'
 export * from './context'
 export * from './doctype'
 export * from './utils/doctype-utils'
-export * from './utils/ipfs-utils'
