+++ packages/ceramic-core/src/__tests__/ceramic-api.test.ts
@@ -3,11 +3,38 @@ import IdentityWallet from 'identity-wallet'
 import Ipfs from 'ipfs'
 import tmp from 'tmp-promise'
 import { TileDoctype, TileParams } from "@ceramicnetwork/ceramic-doctype-tile"
-import { AnchorStatus, IpfsUtils, DoctypeUtils } from "@ceramicnetwork/ceramic-common"
+import { AnchorStatus, DoctypeUtils } from "@ceramicnetwork/ceramic-common"
+
+import dagJose from 'dag-jose'
+import basicsImport from 'multiformats/cjs/src/basics-import.js'
+import legacy from 'multiformats/cjs/src/legacy.js'
 
 jest.mock('../store/level-state-store')
 
 const seed = '0x5872d6e0ae7347b72c9216db218ebbb9d9d0ae7ab818ead3557e8e78bf944184'
+
+/**
+ * Create an IPFS instance
+ * @param overrideConfig - IFPS config for override
+ */
+const createIPFS =(overrideConfig: object = {}): Promise<any> => {
+  basicsImport.multicodec.add(dagJose)
+  const format = legacy(basicsImport, dagJose.name)
+
+  const config = {
+    ipld: { formats: [format] },
+    libp2p: {
+      config: {
+        dht: {
+          enabled: true
+        }
+      }
+    }
+  }
+
+  Object.assign(config, overrideConfig)
+  return Ipfs.create(config)
+}
 
 describe('Ceramic API', () => {
   jest.setTimeout(15000)
@@ -42,7 +69,7 @@ describe('Ceramic API', () => {
   beforeAll(async () => {
     tmpFolder = await tmp.dir({ unsafeCleanup: true })
 
-    ipfs = await IpfsUtils.createIPFS({
+    ipfs = await createIPFS({
       repo: `${tmpFolder.path}/ipfs${9}/`,
       config: {
         Addresses: { Swarm: [ `/ip4/127.0.0.1/tcp/${4013}` ] },
